requirements.txt
flask
psycopg2
python-dotenv
jwt